requirements.txt
attrs==21.4.0
black==22.1.0
click==8.0.3
iniconfig==1.1.1
mypy-extensions==0.4.3
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
PyYAML==6.0
tomli==2.0.1
typing-extensions==4.1.1
